+++ pkg/executor/handler.go
@@ -93,6 +93,11 @@ var handlerList = map[components.ResourceType]map[actions.ActionType]actionHandl
 		actions.ActionDisable: serviceAction,
 		actions.ActionCleanup: cleanupBuildArtifact,
 	},
+	components.ResourceTypeAppFile: {
+		actions.ActionInstall: installOrUpdateFile,
+		actions.ActionUpdate:  installOrUpdateFile,
+		actions.ActionRemove:  removeFile,
+	},
 	components.ResourceTypeFile: {
 		actions.ActionInstall: installOrUpdateFile,
 		actions.ActionUpdate:  installOrUpdateFile,